First commit for PseudoHeaderGen.
This Clang-based tool mirrors a set of headers but writes declarations in a commentless "standard" form, so commenting a line is removing a declaration. Those pseudo-headers are meant to be used by another upcoming tool, PH2LLVMBitcode, that will generate LLVM bitcode files containing the declarations for OGRE, Bullet, cAudio, etc.'s classes and functions, but only the ones which are whitelisted in the pseudo-headers.