Fix compilation, sticking with C++ for now.
I'll probably switch to D in the future, but I'll take my time to figure out
which parser generator to use, the best way to create bindings for LLVM, etc.
Also I'll try to write an auto-completion plugin for KDevelop. There are already
good tools for D in Eclipse, but Eclipse doesn't seem to work nicely with CMake.
Maybe also an interface generator between C++ and D.